+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Next Release
 
-- **[NEW]** Added `rename` command, which changes the remote URL and moves the clone accordingly.
+- **[NEW]** Added `rename` command, which changes the remote URL and moves the clone accordingly
+- **[IMPROVED]** The `mv` command now includes remote information when prompting to choose a directory
 
 ## 0.6.0 (2017-04-01)